+++ app/models/group_bank_account.rb
@@ -1,8 +1,18 @@
 class GroupBankAccount < BankAccount
   belongs_to :group, required: true
-
   delegate :name, to: :group
-  alias_method :owner, :group
+
+  def transactions
+    outgoing_transactions.union(incoming_transactions)
+      .includes(from_account: :group, to_account: :group)
+      .latest
+  end
+
+  def items
+    outgoing_items.union(incoming_items)
+      .includes(from_account: :group, to_account: :group)
+      .latest
+  end
 
   def display_name
     "#{self.group.name} | #{self.chapter.name}"