+++ app/policies/personal_bank_account_policy.rb
@@ -1,0 +1,21 @@
+class PersonalBankAccountPolicy < ApplicationPolicy
+
+  def has_control?
+    record.owner == @character
+  end
+
+  class Scope
+    attr_reader :user, :character, :scope
+
+    def initialize(context, scope)
+      @user = context.user
+      @character = context.character
+      @scope = scope
+    end
+
+    def resolve
+      scope.where(owner: @character)
+    end
+  end
+
+end