+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 (Lol, it tries to, but SemVer is hard. Especially given this project's constraints.)
 
 ## [Unreleased](#) - ????-??-??
+### Removed
+- Rake DB tasks in favor of the shell script from 2.2.0 / 460803cb
 
 ## [2.4.0](v.2.4.0) - 2017-07-04
 ### Added