+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 (Lol, it tries to, but SemVer is hard. Especially given this project's constraints.)
 
 ## [Unreleased](#) - ????-??-??
+### Changed
+- Partially fixed BankAccounts 500-ing (still have some Policy issues)
+
 ### Removed
 - Rake DB tasks in favor of the shell script from 2.2.0 / 460803cb